channelBox receive socket props and no functions
